requirements.txt
django==3.*
djangorestframework==3.12.*
django-extensions==3.1.*
djangorestframework-simplejwt==4.6.*
django-cors-headers==3.5.*
psycopg2==2.8.*
Faker==5.*
pytest-django==4.*
gunicorn==20.0.4
python-dotenv==0.15.0
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
oauth2client==4.1.3
pyOpenSSL==20.0.1
django-debug-toolbar==3.2.*
django-filter==2.4.*
celery==5.0.*
redis==3.5
drf-spectacular==0.13.*
openpyxl==3.0.6
